christmas_tree/rpi/requirements.txt
rpi_ws281x
adafruit-circuitpython-neopixel
adafruit-blinka
Flask
